Broke apart and reorganized much of astropy_helpers.setup_helpers. Most of the code related to implementing individual distutils commands has been moved into an astropy_helpers.commands/ subpackage, with a module devoted to each command.  Some of these modules are very short and seem a bit silly, but this organization is simply to match the organization of distutils and setuptools.  It also makes it easy to find where a given command is implemented.
Also moved most of the functions related to introspecting distutils into
a new distutils_helpers module.
