+++ astropy_helpers/version_helpers.py
@@ -29,7 +29,7 @@ import sys
 from distutils import log
 
 from . import git_helpers
-from .setup_helpers import is_distutils_display_option, get_pkg_version_module
+from .distutils_helpers import is_distutils_display_option
 from .utils import invalidate_caches
 
 PY3 = sys.version_info[0] == 3
@@ -203,3 +203,25 @@ def generate_version_py(packagename, version, release=None, debug=None,
 
         if version_module:
             imp.reload(version_module)
+
+
+def get_pkg_version_module(packagename, fromlist=None):
+    """Returns the package's .version module generated by
+    `astropy_helpers.version_helpers.generate_version_py`.  Raises an
+    ImportError if the version module is not found.
+
+    If ``fromlist`` is an iterable, return a tuple of the members of the
+    version module corresponding to the member names given in ``fromlist``.
+    Raises an `AttributeError` if any of these module members are not found.
+    """
+
+    if not fromlist:
+        # Due to a historical quirk of Python's import implementation,
+        # __import__ will not return submodules of a package if 'fromlist' is
+        # empty.
+        # TODO: For Python 3.1 and up it may be preferable to use importlib
+        # instead of the __import__ builtin
+        return __import__(packagename + '.version', fromlist=[''])
+    else:
+        mod = __import__(packagename + '.version', fromlist=fromlist)
+        return tuple(getattr(mod, member) for member in fromlist)
